main

+++ cogs/check_questions.py
@@ -14,11 +14,12 @@ class check_questions(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, msg):
-        if self.bot.send_message_allowed == False:
+        if self.bot.send_message_allowed == False and msg.author != self.bot.user:
             await msg.delete()
-        if msg.author == self.bot.user:
             return
         if msg.channel != self.channel:
+            return
+        if msg.author.bot:
             return
         
         await self.check_answer(msg, msg.author, self.answer)
@@ -28,7 +29,6 @@ class check_questions(commands.Cog):
         answer = answer3.replace(" ", "")
         msg_content = msg.content.lower().strip()
         if fuzz.ratio(answer, msg_content) >= 80 and len(answer.strip()) >= 5 or fuzz.ratio(answer, msg_content) >= 75 and len(answer.strip()) >= 10 or fuzz.ratio(answer, msg_content) >= 70 and len(answer.strip()) >= 15 or fuzz.ratio(answer, msg_content) >= 65 and len(answer.strip()) >= 20 or fuzz.ratio(answer, msg_content) >= 60 and len(answer.strip()) >= 25 or answer == msg_content:
-            self.bot.send_message_allowed = False
             await msg.add_reaction("✅")
             punkte = self.bot.points_author if user == self.author else self.bot.points_user
             if self.difficulty == "EASY":
@@ -45,6 +45,7 @@ class check_questions(commands.Cog):
             
             embed_correct_answer = dc.Embed(title="✅ Frage beantwortet ✅", description=f"Glückwunsch {user.mention}, deine Antwort ist richtig!\n\nDie richtige Antwort lautet: **{answer1}**\n\n **Aktueller Punktestand**\n {self.author.mention}: {self.bot.points_author} \n {self.user.mention}: {self.bot.points_user}\n", color=0x00D166)
             await msg.channel.send(embed=embed_correct_answer)
+            self.bot.send_message_allowed = False
         else:
             self.bot.send_message_allowed = True
             await msg.add_reaction("❌")
